+++ src/apps/weblib/cinterface.h
@@ -1,6 +1,8 @@
 #ifndef LIB_CINTERFACE_H
 #define LIB_CINTERFACE_H
 
+#include <cstdint>
+
 extern "C" {
 
 extern void vizzu_init();
@@ -14,7 +16,7 @@ extern void vizzu_keyPress(int key, bool ctrl, bool alt, bool shift);
 extern void vizzu_setLogging(bool enable);
 extern void
 vizzu_update(double width, double height, int renderControl);
-extern const char *vizzu_errorMessage(int exceptionPtr);
+extern const char *vizzu_errorMessage(intptr_t exceptionPtr);
 extern const char *vizzu_version();
 
 extern void data_addDimension(const char *name,
